requirements.txt
anyio==3.6.2
bcrypt==4.0.1
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-log==0.4.0
colorama==0.4.6
cryptography==3.4.8
deprecation==2.1.0
dnspython==2.2.1
docutils==0.19
dotty-dict==1.3.1
email-validator==1.3.0
fastapi==0.86.0
fastapi-jwt-auth==0.5.0
gitdb==4.0.9
GitPython==3.1.29
gotrue==0.5.4
h11==0.12.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
idna==3.4
importlib-metadata==5.0.0
invoke==1.7.3
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
keyring==23.11.0
MarkupSafe==2.1.1
more-itertools==9.0.0
orjson==3.8.1
packaging==21.3
passlib==1.7.4
pkginfo==1.8.3
postgrest-py==0.10.3
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-gitlab==3.11.0
python-multipart==0.0.5
python-semantic-release==7.32.1
pywin32-ctypes==0.2.0
PyYAML==6.0
readme-renderer==37.3
realtime==0.0.5
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.20.4
storage3==0.3.5
supabase==0.7.1
supafunc==0.2.2
tomlkit==0.11.6
tqdm==4.64.1
twine==3.8.0
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.18.1
webencodings==0.5.1
websockets==10.4
zipp==3.10.0
